+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/controllers/FileController.java
@@ -21,16 +21,22 @@ package com.onlyoffice.integration.controllers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.google.gson.reflect.TypeToken;
 import com.onlyoffice.integration.documentserver.managers.history.HistoryManager;
 import com.onlyoffice.integration.documentserver.storage.FileStorageMutator;
 import com.onlyoffice.integration.documentserver.storage.FileStoragePathBuilder;
 import com.onlyoffice.integration.dto.Converter;
+import com.onlyoffice.integration.dto.FormatsList;
 import com.onlyoffice.integration.dto.Reference;
+import com.onlyoffice.integration.dto.RefreshConfig;
 import com.onlyoffice.integration.dto.Rename;
 import com.onlyoffice.integration.dto.Restore;
 import com.onlyoffice.integration.dto.SaveAs;
+import com.onlyoffice.integration.dto.RefreshConfig.Document;
+import com.onlyoffice.integration.dto.RefreshConfig.EditorConfig;
 import com.onlyoffice.integration.entities.User;
 import com.onlyoffice.integration.sdk.manager.DocumentManager;
+import com.onlyoffice.integration.sdk.service.ConfigService;
 import com.onlyoffice.integration.services.UserServices;
 
 import com.onlyoffice.manager.request.RequestManager;
@@ -48,7 +54,7 @@ import com.onlyoffice.model.documenteditor.config.document.ReferenceData;
 import com.onlyoffice.service.command.CommandService;
 import com.onlyoffice.service.convert.ConvertService;
 import com.onlyoffice.service.documenteditor.callback.CallbackService;
-import org.apache.http.HttpEntity;
+import org.apache.hc.core5.http.HttpEntity;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
@@ -121,6 +127,8 @@ public class FileController {
     @Autowired
     private CommandService commandService;
     @Autowired
+    private ConfigService configService;
+    @Autowired
     private UrlManager urlManager;
 
     // create user metadata
@@ -133,7 +141,7 @@ public class FileController {
             storageMutator.createMeta(fullFileName,  // create meta information with the user ID and name specified
                     String.valueOf(user.getId()), user.getName());
         }
-        return "{ \"filename\": \"" + fullFileName + "\", \"documentType\": \"" + documentType + "\" }";
+        return "{ \"filename\": \"" + fullFileName + "\", \"documentType\": \"" + documentType + "\",\"percent\":100}";
     }
 
     // download data from the specified file
@@ -237,10 +245,11 @@ public class FileController {
         String filePass = body.getFilePass() != null ? body.getFilePass() : null;
         // get an auto-conversion extension from the request body or set it to the ooxml extension
         String conversionExtension = body.getFileExt() != null ? body.getFileExt() : "ooxml";
+        Boolean keepOriginal = body.getKeepOriginal() != null ? body.getKeepOriginal() : false;
 
         try {
             // check if the file with such an extension can be converted
-            if (documentManager.getDefaultConvertExtension(fileName) != null) {
+            if (documentManager.getDefaultConvertExtension(fileName) != null || body.getFileExt() != null) {
                 ConvertRequest convertRequest = ConvertRequest.builder()
                         .password(filePass)
                         .outputtype(conversionExtension)
@@ -256,6 +265,13 @@ public class FileController {
 
                 String newFileUri = convertResponse.getFileUrl();
                 String newFileType = convertResponse.getFileType();
+
+                if (!new FormatsList(documentManager.getFormats()).getFormats().stream().anyMatch(
+                        f -> newFileType.equals(f.getName()) && f.getType() != null)
+                    ) {
+                        return "{ \"percent\" : \"100\", \"filename\" : \"" + newFileUri
+                                     + "\", \"error\":\"FileTypeIsNotSupported\"}";
+                }
 
                 /* get a file name of an internal file extension with an index if the file
                  with such a name already exists */
@@ -274,7 +290,9 @@ public class FileController {
 
 
                         // remove source file
-                        storageMutator.deleteFile(oldFileName);
+                        if (!keepOriginal) {
+                            storageMutator.deleteFile(oldFileName);
+                        }
 
                         // create the converted file with input stream
                         storageMutator.createFile(Path.of(storagePathBuilder.getFileLocation(correctedName)), stream);
@@ -681,6 +699,42 @@ public class FileController {
             responseBody.put("error", error.getMessage());
             responseBody.put("success", false);
             return responseBody.toJSONString();
+        }
+    }
+
+    @GetMapping("/config")
+    public ResponseEntity<RefreshConfig> config(@RequestParam("fileName") final String fileName,
+                            @RequestParam("permissions") final String permissions,
+                            @CookieValue(value = "uid", required = false) final String uid) {
+        try {
+            if (!new File(storagePathBuilder.getFileLocation(fileName)).exists()) {
+                throw(new Exception("File not found"));
+            }
+
+            Gson gson = new GsonBuilder().disableHtmlEscaping().create();
+
+            Document document = new Document();
+            document.setTitle(fileName);
+            document.setKey(documentManager.getDocumentKey(fileName, false));
+            document.setUrl(urlManager.getFileUrl(fileName));
+            document.setReferenceData(configService.getReferenceData(fileName));
+            document.setPermissions(gson.fromJson(permissions, new TypeToken<Map<String, Object>>() { }.getType()));
+
+            EditorConfig editorConfig = new EditorConfig();
+            editorConfig.setCallbackUrl(urlManager.getCallbackUrl(fileName));
+
+            RefreshConfig config = new RefreshConfig();
+            config.setDocument(document);
+            config.setEditorConfig(editorConfig);
+            if (settingsManager.isSecurityEnabled()) {
+                config.setToken(jwtManager.createToken(config));
+            }
+
+            return ResponseEntity.ok(config);
+        } catch (Exception e) {
+            RefreshConfig error = new RefreshConfig();
+            error.setError(e.getMessage());
+            return ResponseEntity.internalServerError().body(error);
         }
     }
 }