+++ web/documentserver-example/php-laravel/app/UseCases/Docs/Conversion/ConvertCommand.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * (c) Copyright Ascensio System SIA 2024
  *
@@ -34,7 +35,7 @@ class ConvertCommand
     {
         $format = $this->formatRepository->find($request->fileType);
 
-        if (! $format->convertible()) {
+        if (! $format->convertible() && $request->outputType == 'ooxml') {
             throw new Exception("The format $request->fileType is not convertible.");
         }