+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
-- editing by default
-- change reference source
-- filename in editor page title
-- close editor
+- nodejs: support vsdx in visio editor
+- nodejs: support pages, numbers, key formats
+- golang: new integration example
+- golang: upload files to the server
+- golang: create blank files and files with sample content
+- golang: edit uploaded files in onlyoffice editor
+- golang: delete files on the server
+- golang: show files history
 - php-laravel: new integration example
 - php-laravel: upload files to the server
 - php-laravel: create blank files and files with sample content
@@ -12,6 +16,14 @@
 - php-laravel: create, edit, and submit pdf forms
 - php-laravel: show forgotten files on a seperate page
 - php-laravel: fetch files
+- refresh config
+- on uploading xml convert to supported type only
+- converting function on index page
+- editing by default
+- change reference source
+- filename in editor page title
+- close editor
+- sq-AL skin language
 
 ## 1.11.0
 - he-IL skin language